requirements.txt
litellm>=1.18.3
numpy>=1.24.3
pytest>=7.4.0
psycopg2-binary>=2.9.9
tqdm>=4.66.1
aiohttp>=3.9.1
openai>=1.3.8 